feat(banner): improve banner text capabilities
This changes the banner to support formatted numbers in the banners texts.

As AppEngine has no support for locale und string.format() does not support
formatting number with dots as thousand separators we add `util.py` for
formatting numbers in the formats needed.

The replacing of the formatted numbers in the config has also been generalized.
All string properties of the config json object are walked and updated.
See `test_config_replacement`

Keys which are replaced are:

 * `donated`: amount alreaday donated
 * `unlocked`: amount unlocked in current stretch
 * `percent`: completion of current goal
 * `clicks`: number of clicks in current stretch
 * `increment`: amount donated per click

The behaviour of the banner count has been updated to contain the **global**
clickcount and progress and not that of the current domain. There is simple no
"progress" for a single domain.